+++ Tracer/include/shapes/sphere.hpp
@@ -3,6 +3,8 @@
 
 #include <memory>
 
+#include "float.hpp"
+
 #include "hittable.hpp"
 #include "hitRecord.hpp"
 
@@ -15,15 +17,15 @@ class Sphere : public HittableObject {
 	public:
 		// No default constructor because this class is not default-constructible.
 
-        Sphere(point3 center, double radius, std::shared_ptr<Material> material) :
+        Sphere(point3 center, Float radius, std::shared_ptr<Material> material) :
 			m_center(center), m_radius(radius), m_material(material) {}
 
         virtual bool hit(
-			const ray& r, double tMin, double tMax, HitRecord& rec) const override;
+			const ray& r, Float tMin, Float tMax, HitRecord& rec) const override;
 
     private:
         point3 m_center;
-        double m_radius;
+        Float m_radius;
 		std::shared_ptr<Material> m_material;
 };
