+++ Tracer/include/shared/hittable.hpp
@@ -1,12 +1,14 @@
 #ifndef HITTABLE_HPP
 #define HITTABLE_HPP
 
+#include "float.hpp"
+
 #include "hitRecord.hpp"
 #include "ray.hpp"
 
 class HittableObject {
 	public:
-		virtual bool hit(const ray& r, double tMin, double tMax, HitRecord& rec) const = 0;
+		virtual bool hit(const ray& r, Float tMin, Float tMax, HitRecord& rec) const = 0;
 };
 
 #endif // HITTABLE_HPP
