+++ Tracer/include/shared/hittableList.hpp
@@ -21,7 +21,7 @@ class HittableList : public HittableObject {
 			m_objects.push_back(object);
 		}
 
-        virtual bool hit(const ray& r, double tMin, double tMax, HitRecord& rec) const override;
+        virtual bool hit(const ray& r, Float tMin, Float tMax, HitRecord& rec) const override;
 
     private:
         std::vector<std::shared_ptr<HittableObject>> m_objects;
